# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOC Analyzer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Upload",
        "operationId": "create_upload_api_uploads_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_api_uploads_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "List Uploads",
        "description": "Newest first. Summary fields only (no bad-line samples).",
        "operationId": "list_uploads_api_uploads_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UploadOut"
                  },
                  "title": "Response List Uploads Api Uploads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/uploads/{upload_id}/summary": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Summary",
        "description": "Dashboard data: stats, timeline and top-10 lists (one request for the whole page).",
        "operationId": "get_summary_api_uploads__upload_id__summary_get",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/uploads/{upload_id}/incidents": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "List Incidents",
        "description": "Incidents worst first, each with its findings. `min_priority=medium` hides the low ones.",
        "operationId": "list_incidents_api_uploads__upload_id__incidents_get",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Upload Id"
            }
          },
          {
            "name": "min_priority",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "critical",
                "high",
                "medium",
                "low"
              ],
              "type": "string",
              "default": "low",
              "title": "Min Priority"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_upload_api_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone",
            "default": "UTC"
          },
          "format": {
            "type": "string",
            "enum": [
              "auto",
              "csv",
              "json"
            ],
            "title": "Format",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_api_uploads_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Timeline": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "points"
        ],
        "title": "Timeline"
      },
...
      "UploadCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Count"
          },
          "bad_line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Line Count"
          },
          "previous_uploads": {
            "type": "integer",
            "title": "Previous Uploads"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "format",
          "log_timezone",
          "status",
          "progress",
          "size_bytes",
          "sha256",
          "uploaded_by",
          "created_at",
          "error",
          "line_count",
          "bad_line_count",
          "previous_uploads"
        ],
        "title": "UploadCreated"
      },
...
      "UploadOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "log_timezone": {
            "type": "string",
            "title": "Log Timezone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Count"
          },
          "bad_line_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bad Line Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "format",
          "log_timezone",
          "status",
          "progress",
          "size_bytes",
          "sha256",
          "uploaded_by",
          "created_at",
          "error",
          "line_count",
          "bad_line_count"
        ],
        "title": "UploadOut"
      },
      "UploadSummaryOut": {
        "properties": {
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          },
          "stats": {
            "$ref": "#/components/schemas/SummaryStats"
          },
          "timeline": {
            "$ref": "#/components/schemas/Timeline"
          },
          "top": {
            "$ref": "#/components/schemas/TopLists"
          },
          "computed_at": {
            "type": "string",
            "title": "Computed At"
          },
          "narrative": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NarrativeOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "stats",
          "timeline",
          "top",
          "computed_at",
          "narrative"
        ],
        "title": "UploadSummaryOut"
      },
...
    }
  }
}